demandes

+++ routes/demande.js
@@ -9,7 +9,7 @@ const { checkBody } = require('../modules/checkBody');
 //créer une nouvelle demande
 
 router.post('/', async (req, res) => {
-    if (!checkBody(req.body, ['token', 'content'])) {
+    if (!checkBody(req.body, ['token'])) {
       res.json({ result: false, error: 'Missing or empty fields' });
       return;
     }
@@ -43,7 +43,7 @@ router.get('/:id', async (req, res) => {
     try {
         const demande = await Demande.findById(req.params.id);
         if (!demande) {
-            return res.json({ error: 'Demande not found' });
+            return res.json({ error: 'Request not found' });
         }
         res.json(demande);
     } catch (err) {
@@ -59,7 +59,7 @@ router.get('/', async (req, res) => {
     try {
         const demandes = await Demande.find();
         if (!demandes) {
-            return res.json({ error: 'Demandes not found' });
+            return res.json({ error: 'Requests not found' });
         }
         res.json(demandes);
     } catch (err) {
@@ -71,18 +71,29 @@ router.get('/', async (req, res) => {
 
 //mettre à jour une demande
 router.put('/:id', async (req, res) => {
-    const { statut, message, type } = req.body;
-    const updateDemande = {
-        statut,
-        message,
-        type,
-        dateMAJ: Date.now(),
-    };
+    if (!checkBody(req.body, ['token'])) {
+        res.json({ result: false, error: 'Missing or empty fields' });
+        return;
+    }
 
     try {
-        const updatedDoc = await Demande.findByIdAndUpdate(req.params.id, updateDemande, { new: true });
+        const user = await User.findOne({ token: req.body.token });
+        if (!user) {
+            res.json({ result: false, error: 'User not found' });
+            return;
+        }
+
+        const { statut, message, type } = req.body;
+        const updateDemande = {
+            statut,
+            message,
+            type,
+            dateMAJ: Date.now(),
+        };
+
+        const updatedDoc = await Demande.findByIdAndUpdate(req.params.id, updateDemande, { new: true });  //met à jour la demande en BDD et retourne une réponse
         if (!updatedDoc) {
-            return res.json({ error: 'Demande not found' });
+            return res.json({ error: 'Request not found' });
         }
         res.json(updatedDoc);
     } catch (err) {
@@ -94,14 +105,25 @@ router.put('/:id', async (req, res) => {
 
 // supprimer une demande
 router.delete('/:id', async (req, res) => {
+    if (!checkBody(req.body, ['token'])) {
+        res.json({ result: false, error: 'Missing or empty fields' });
+        return;
+    }
+
     try {
+        const user = await User.findOne({ token: req.body.token });
+        if (!user) {
+            res.json({ result: false, error: 'User not found' });
+            return;
+        }
+    
       const deletedDemande = await Demande.findByIdAndDelete(req.params.id);
   
       if (!deletedDemande) {
         return res.json({ error: 'Demande not found' });
       }
   
-      res.json({ message: 'Demande deleted successfully' });
+      res.json({ message: 'request deleted successfully' });
     } catch (err) {
       res.json({ error: err.message });
     }
